# swagger.json
{
   "swagger": "2.0",
   "info": {
     "title": "API - User and Doctor management",
     "description": "base_url: https://api.homedoc.fr",
     "version": "1.0"
   },
   "produces": ["application/json"],
  "consumes": ["application/json"],
  "paths": {
     "/createUser": {
       "post": {
         "tags": ["user"],
         "x-swagger-router-controller": "bar",
         "operationId": "1",
         "description": "[Login 123](https://www.google.com)",
         "parameters": [{
           "name": "body",
           "in": "body",
           "required": true,
           "schema": {
             "$ref": "#/definitions/createUser"
           }
         }
         ],
         "responses": {
           "200" : {
             "description": "Success"
           }
         }
       }
     },
     "/login": {
		"post": {
          "tags": ["user"],
          "x-swagger-router-controller": "bar",
		  "operationId": "2",
		  "description": "",
		  "responses": {},
		  "parameters": [{
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/login"
            },
            "name": "body"
          }
          ]
		}
	  },
    "/editUser": {
      "post": {
        "tags": ["user"],
        "x-swagger-router-controller": "bar",
        "operationId": "3",
        "description": "",
        "responses": {},
        "parameters": [{
          "in": "body",
          "required": true,
          "schema": {
            "$ref": "#/definitions/editUser"
          },
          "name": "body"
        }
        ]
      }
    },
    "/getUser": {
      "post": {
        "tags": ["user"],
        "x-swagger-router-controller": "bar",
        "operationId": "4",
        "description": "",
        "responses": {},
        "parameters": [{
          "in": "body",
          "required": true,
          "schema": {
            "$ref": "#/definitions/getUser"
          },
          "name": "body"
        }
        ]
      }
    },
    "/logout": {
      "post": {
        "tags": ["user"],
        "x-swagger-router-controller": "bar",
        "operationId": "5",
        "description": "",
        "responses": {},
        "parameters": [{
          "in": "body",
          "required": true,
          "schema": {
            "$ref": "#/definitions/logout"
          },
          "name": "body"
        }
        ]
      }
    },
    "/usersList": {
      "post": {
        "tags": ["user"],
        "x-swagger-router-controller": "bar",
        "operationId": "6",
        "description": "",
        "responses": {},
        "parameters": [{
          "in": "body",
          "required": false,
          "schema": {
          },
          "name": "body"
        }
        ]
      }
    },
    "/changePassword": {
      "post": {
        "tags": ["user"],
        "x-swagger-router-controller": "bar",
        "operationId": "7",
        "description": "",
        "responses": {},
        "parameters": [{
          "in": "body",
          "required": true,
          "schema": {
            "$ref": "#/definitions/changePassword"
          },
          "name": "body"
        }
        ]
      }
    },
    "/loginDoctor": {
      "post": {
        "tags": ["doctor"],
        "x-swagger-router-controller": "bar",
        "operationId": "8",
        "description": "",
        "responses": {},
        "parameters": [{
          "in": "body",
          "required": true,
          "schema": {
            "$ref": "#/definitions/login"
          },
          "name": "body"
        }
        ]
      }
    },
    "/logoutDoctor": {
      "post": {
        "tags": ["doctor"],
        "x-swagger-router-controller": "bar",
        "operationId": "9",
        "description": "",
        "responses": {},
        "parameters": [{
          "in": "body",
          "required": true,
          "schema": {
            "$ref": "#/definitions/logout"
          },
          "name": "body"
        }
        ]
      }
    },
    "/createDoctor": {
      "post": {
        "tags": ["doctor"],
        "x-swagger-router-controller": "bar",
        "operationId": "10",
        "description": "[Login 123](https://www.google.com)",
        "parameters": [{
          "name": "body",
          "in": "body",
          "required": true,
          "schema": {
            "$ref": "#/definitions/createUser"
          }
        }
        ],
        "responses": {
          "200" : {
            "description": "Success"
          }
        }
      }
    },
    "/getDoctor": {
      "post": {
        "tags": ["doctor"],
        "x-swagger-router-controller": "bar",
        "operationId": "11",
        "description": "",
        "responses": {},
        "parameters": [{
          "in": "body",
          "required": true,
          "schema": {
            "$ref": "#/definitions/getUser"
          },
          "name": "body"
        }
        ]
      }
    },
    "/editDoctor": {
      "post": {
        "tags": ["doctor"],
        "x-swagger-router-controller": "bar",
        "operationId": "12",
        "description": "",
        "responses": {},
        "parameters": [{
          "in": "body",
          "required": true,
          "schema": {
            "$ref": "#/definitions/editDoctor"
          },
          "name": "body"
        }
        ]
      }
    },
    "/changePasswordDoctor": {
      "post": {
        "tags": ["doctor"],
        "x-swagger-router-controller": "bar",
        "operationId": "13",
        "description": "",
        "responses": {},
        "parameters": [{
          "in": "body",
          "required": true,
          "schema": {
            "$ref": "#/definitions/changePassword"
          },
          "name": "body"
        }
        ]
      }
    }
  },
  "definitions": {
    "login" : {
      "properties" : {
        "email" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        }
      }
    },
    "createUser" : {
      "properties" : {
        "email" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        },
        "firstname" : {
          "type" : "string"
        },
        "lastname" : {
          "type" : "string"
        },
        "dateOfBirth" : {
          "type" : "string"
        }
      }
    },
    "editUser" : {
      "properties" : {
        "firstname" : {
          "type" : "string"
        },
        "lastname" : {
          "type" : "string"
        },
        "dateOfBirth" : {
          "type" : "string"
        },
        "height" : {
          "type" : "number",
          "format" : "double"
        },
        "weight" : {
          "type" : "number",
          "format" : "double"
        },
        "medicalHistory" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "gender" : {
          "type" : "string"
        },
        "allergies" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "city" : {
          "type" : "string"
        },
        "token" : {
          "type" : "string"
        }
      }
    },
    "logout" : {
      "properties" : {
        "token" : {
          "type" : "string"
        }
      }
    },
    "getUser" : {
      "properties" : {
        "token" : {
          "type" : "string"
        }
      }
    },
    "changePassword" : {
      "properties" : {
        "token" : {
          "type" : "string"
        },
        "oldPassword" : {
          "type" : "string"
        },
        "newPassword" : {
          "type" : "string"
        }
      }
    },
    "editDoctor" : {
      "properties" : {
        "firstname" : {
          "type" : "string"
        },
        "lastname" : {
          "type" : "string"
        },
        "dateOfBirth" : {
          "type" : "string"
        },
        "city" : {
          "type" : "string"
        },
        "token" : {
          "type" : "string"
        }
      }
    }
  }
}